+++ src/lib.rs
@@ -17,7 +17,7 @@ pub use multimap_table::{
 pub use table::{RangeIter, ReadOnlyTable, ReadableTable, Table};
 pub use transactions::{DatabaseStats, Durability, ReadTransaction, WriteTransaction};
 pub use tree_store::{AccessGuard, Savepoint};
-pub use types::{RedbKey, RedbValue};
+pub use types::{RedbKey, RedbValue, TypeName};
 
 type Result<T = (), E = Error> = std::result::Result<T, E>;
