+++ src/table.rs
@@ -341,83 +341,3 @@ impl<'a, K: RedbKey + ?Sized + 'a, V: RedbValue + ?Sized + 'a> DoubleEndedIterat
         }
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::types::{RedbKey, RedbValue, Sealed, TypeName};
-    use crate::{Database, ReadableTable, TableDefinition};
-    use std::cmp::Ordering;
-    use tempfile::NamedTempFile;
-
-    #[test]
-    fn custom_ordering() {
-        #[derive(Debug)]
-        struct ReverseKey(Vec<u8>);
-
-        impl RedbValue for ReverseKey {
-            type SelfType<'a> = ReverseKey
-            where
-                Self: 'a;
-            type AsBytes<'a> = &'a [u8]
-            where
-                Self: 'a;
-
-            fn fixed_width() -> Option<usize> {
-                None
-            }
-
-            fn from_bytes<'a>(data: &'a [u8]) -> ReverseKey
-            where
-                Self: 'a,
-            {
-                ReverseKey(data.to_vec())
-            }
-
-            fn as_bytes<'a, 'b: 'a>(value: &'a Self::SelfType<'b>) -> &'a [u8]
-            where
-                Self: 'a,
-                Self: 'b,
-            {
-                &value.0
-            }
-
-            fn type_name() -> TypeName {
-                TypeName::new("test::ReverseKey")
-            }
-        }
-
-        impl Sealed for ReverseKey {}
-
-        impl RedbKey for ReverseKey {
-            fn compare(data1: &[u8], data2: &[u8]) -> Ordering {
-                data2.cmp(data1)
-            }
-        }
-
-        let definition: TableDefinition<ReverseKey, &str> = TableDefinition::new("x");
-
-        let tmpfile: NamedTempFile = NamedTempFile::new().unwrap();
-        let db = Database::create(tmpfile.path()).unwrap();
-        let write_txn = db.begin_write().unwrap();
-        {
-            let mut table = write_txn.open_table(definition).unwrap();
-            for i in 0..10u8 {
-                let key = vec![i];
-                table.insert(&ReverseKey(key), "value").unwrap();
-            }
-        }
-        write_txn.commit().unwrap();
-
-        let read_txn = db.begin_read().unwrap();
-        let table = read_txn.open_table(definition).unwrap();
-        let start = ReverseKey(vec![7u8]); // ReverseKey is used, so 7 < 3
-        let end = ReverseKey(vec![3u8]);
-        let mut iter = table.range(start..=end).unwrap();
-        for i in (3..=7u8).rev() {
-            let (key, value) = iter.next().unwrap();
-            assert_eq!(&[i], key.value().0.as_slice());
-            assert_eq!("value", value.value());
-        }
-        assert!(iter.next().is_none());
-    }
-}
